// frontend/src/app/api/model/adminUserListItemDto.ts
/**
 * CoreData API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RoleDto } from './roleDto';
import { StaffMiniDto } from './staffMiniDto';
import { StudentMiniDto } from './studentMiniDto';


export interface AdminUserListItemDto { 
    id?: string;
    email?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    isActive?: boolean;
    permissions?: number;
    roles?: Array<RoleDto> | null;
    student?: StudentMiniDto;
    staff?: StaffMiniDto;
}